% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGoogleAnalyticsArtifacts.R
\name{createGoogleAnalyticsArtifacts}
\alias{createGoogleAnalyticsArtifacts}
\title{Create a google analytics artifact by reading in all existing raw data generated
by Python/createGoogleAnalyticsArtifacts.py for previous dates, then assessing
dates that must be run and adding those.}
\usage{
createGoogleAnalyticsArtifacts(subdir)
}
\arguments{
\item{subdir}{sub-directory}
}
\description{
Create a google analytics artifact by reading in all existing raw data generated
by Python/createGoogleAnalyticsArtifacts.py for previous dates, then assessing
dates that must be run and adding those.
}
