% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImmuneSpaceRInitsArtifact.R
\name{createImmuneSpaceRInitsArtifact}
\alias{createImmuneSpaceRInitsArtifact}
\title{Create ImmuneSpaceR Inits artifact for use in multiple modules}
\usage{
createImmuneSpaceRInitsArtifact(subdir)
}
\description{
Create ImmuneSpaceR Inits artifact for use in multiple modules
}
