% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParsedLogsArtifact.R
\name{createParsedLogsArtifact}
\alias{createParsedLogsArtifact}
\title{Create artifact from tomcat server logs to use for downstream analysis}
\usage{
createParsedLogsArtifact(subdir)
}
\description{
Create artifact from tomcat server logs to use for downstream analysis
}
