% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPubMedArtifact.R
\name{createPubMedArtifact}
\alias{createPubMedArtifact}
\title{Create artifact from PubMed data for use in ResourcesPage and MonitorIS to describe
publications related to ImmuneSpace}
\usage{
createPubMedArtifact(subdir)
}
\arguments{
\item{subdir}{sub-directory}
}
\description{
Create artifact from PubMed data for use in ResourcesPage and MonitorIS to describe
publications related to ImmuneSpace
}
