% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResourcesPageArtifacts.R
\name{createResourcesPageArtifacts}
\alias{createResourcesPageArtifacts}
\title{Create artifact from tomcat server logs to use for downstream analysis}
\usage{
createResourcesPageArtifacts(subdir)
}
\description{
Create artifact from tomcat server logs to use for downstream analysis
}
