% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSharedArtifacts.R
\name{createSharedArtifacts}
\alias{createSharedArtifacts}
\title{Create ImmuneSpaceR Inits artifact for use in multiple modules}
\usage{
createSharedArtifacts(subdir)
}
\arguments{
\item{subdir}{sub-directory}
}
\description{
Create ImmuneSpaceR Inits artifact for use in multiple modules
}
