% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCurrentRDS}
\alias{getCurrentRDS}
\title{Read in latest RDS object for a given filename}
\usage{
getCurrentRDS(subdir, filename)
}
\arguments{
\item{subdir}{location of artifacts on rsT or rsP}

\item{filename}{filename suffix to use for saving and cleaning up}
}
\description{
Read in latest RDS object for a given filename
}
