% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveAndCleanUp}
\alias{saveAndCleanUp}
\title{Save RDS object and clean up any previous versions}
\usage{
saveAndCleanUp(data, subdir, filename)
}
\arguments{
\item{data}{the data to save}

\item{subdir}{location of artifacts on rsT or rsP}

\item{filename}{filename suffix to use for saving and cleaning up}
}
\description{
Save RDS object and clean up any previous versions
}
