% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{searchStrings}
\alias{searchStrings}
\title{Strings for parsing raw logs data}
\format{
A list of strings \code{modules}, \code{reports},
  \code{rstudio}, and \code{studies}
}
\usage{
searchStrings
}
\description{
Strings to be used in regular expression matching to find specific
types of user interactions within the server logs data
}
\keyword{datasets}
