% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAllArtifacts.R
\name{CreateAllArtifacts}
\alias{CreateAllArtifacts}
\title{Create all artifacts for use in multiple modules}
\usage{
CreateAllArtifacts(subdir = "/share/files/Studies/R_API_resources/")
}
\description{
Create all artifacts for use in multiple modules
}
