% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMonitorISArtifacts.R
\name{createMonitorISArtifacts}
\alias{createMonitorISArtifacts}
\title{Create artifacts for MonitorIS module report using parsed logs}
\usage{
createMonitorISArtifacts(subdir)
}
\description{
Create artifacts for MonitorIS module report using parsed logs
}
